refactor: categoryList
